small fixes

+++ java/src/main/java/org/monetdb/monetdbe/MonetNative.java
@@ -9,6 +9,8 @@ import java.nio.ByteBuffer;
 import java.nio.file.*;
 import java.util.*;
 import java.util.stream.Collectors;
+import java.util.zip.CRC32;
+import java.util.zip.CheckedInputStream;
 
 /**
  * Interface for C native methods in MonetDBe-Java.
@@ -105,7 +107,6 @@ public class MonetNative {
         //Loading within jar
         if ("jar".equalsIgnoreCase(uri.getScheme())) {
             libRoot = FileSystems.newFileSystem(uri, Collections.emptyMap()).getPath("/lib/" + os);
-            //System.out.println("Loading dependencies from within jar: " + libRoot.toString());
         }
         //Loading from file (IDE execution and maven unit tests)
         else {
@@ -116,8 +117,6 @@ public class MonetNative {
             else {
                 libRoot = Paths.get(uri.getPath() + os);
             }
-
-            //System.out.println("Loading dependencies from filesystem: " + libRoot.toString());
         }
         Map<String, List<String>> dependencies = Files.walk(libRoot, 2)
                 .collect(Collectors.groupingBy((path -> path.getParent().getFileName().toString()),
@@ -164,11 +163,30 @@ public class MonetNative {
             throw new IOException("Library " + libName + " could not be found.");
         }
         File loadFile = new java.io.File(System.getProperty("java.io.tmpdir") + "/" + libName);
-        //Clean up file on JVM exit
-        loadFile.deleteOnExit();
         Path tempLibFile = loadFile.toPath();
-        Files.copy(is, tempLibFile, StandardCopyOption.REPLACE_EXISTING);
+        //If the temporary location library already exists, and the checksum matches the library in the jar
+        //No need to copy, only load
+        if (!loadFile.exists() || !libIsUpdated(is,tempLibFile)) {
+            //Clean up file on JVM exit
+            loadFile.deleteOnExit();
+            //Copy library contents in jar to temporary location
+            Files.copy(is, tempLibFile, StandardCopyOption.REPLACE_EXISTING);
+        }
         System.load(tempLibFile.toString());
+    }
+
+    //Compare the checksums from the library in the jar and the library in the temp location
+    //Used to check if the library in the jar needs to be copied, or the old one can be loaded
+    static boolean libIsUpdated(InputStream newLib, Path oldLibPath) throws IOException {
+        InputStream oldLib = Files.newInputStream(oldLibPath);
+        CheckedInputStream newIS = new CheckedInputStream(newLib, new CRC32());
+        CheckedInputStream oldIS = new CheckedInputStream(oldLib, new CRC32());
+        byte[] buffer = new byte[1024];
+        while (newIS.read(buffer, 0, buffer.length) >= 0);
+        long newChecksum = newIS.getChecksum().getValue();
+        while (oldIS.read(buffer, 0, buffer.length) >= 0);
+        long oldChecksum = oldIS.getChecksum().getValue();
+        return newChecksum == oldChecksum;
     }
 
     //Native library API
@@ -193,7 +211,7 @@ public class MonetNative {
      * @param nr_threads     Option for monetdbe_open() library function
      * @return Error message
      */
-    protected static native String monetdbe_open(String dbdir, MonetConnection conn, int sessiontimeout, int querytimeout, int memorylimit, int nr_threads);
+    protected static native String monetdbe_open(String dbdir, MonetConnection conn, int sessiontimeout, int querytimeout, int memorylimit, int nr_threads, String logfile);
 
     /**
      * Open connection to remote connection database.
@@ -211,7 +229,7 @@ public class MonetNative {
      * @param password       Password in the remote database
      * @return Error message
      */
-    protected static native String monetdbe_open(String dbdir, MonetConnection conn, int sessiontimeout, int querytimeout, int memorylimit, int nr_threads, String host, int port, String database, String user, String password);
+    protected static native String monetdbe_open(String dbdir, MonetConnection conn, int sessiontimeout, int querytimeout, int memorylimit, int nr_threads, String host, int port, String database, String user, String password, String logfile);
 
     /**
      * Close the database connection.
